refactor: extract top navigation bar to its own file
